split into engine and game dir
