+++ src/app/engine/model/faction.model.ts
