+++ src/app/engine/model/game-state-store.model.ts
