+++ src/app/engine/model/game-state.model.ts
