+++ src/app/engine/model/piece.model.ts
