+++ src/app/game/dominant-species.constants.ts
