+++ src/app/game/model/element.model.ts
